+++ packages/node.resolver/src/index.ts
@@ -7,27 +7,54 @@ interface NodeVersion {
   lts: false | string
 }
 
+const SEMVER_OPTS = {
+  includePrerelease: true,
+  loose: true,
+}
+
 export async function resolveNodeVersion (
   fetch: FetchFromRegistry,
   versionSpec: string,
   nodeMirrorBaseUrl?: string
 ): Promise<string | null> {
-  const response = await fetch(`${nodeMirrorBaseUrl ?? 'https://nodejs.org/download/release/'}index.json`)
-  const allVersions = (await response.json()) as NodeVersion[]
+  const allVersions = await fetchAllVersions(fetch, nodeMirrorBaseUrl)
   if (versionSpec === 'latest') {
-    return allVersions[0].version.substring(1)
+    return allVersions[0].version
   }
   const { versions, versionRange } = filterVersions(allVersions, versionSpec)
-  const pickedVersion = semver.maxSatisfying(
-    versions.map(({ version }) => version), versionRange, { includePrerelease: true, loose: true })
-  if (!pickedVersion) return null
-  return pickedVersion.substring(1)
+  return semver.maxSatisfying(versions, versionRange, SEMVER_OPTS) ?? null
+}
+
+export async function resolveNodeVersions (
+  fetch: FetchFromRegistry,
+  versionSpec?: string,
+  nodeMirrorBaseUrl?: string
+): Promise<string[]> {
+  const allVersions = await fetchAllVersions(fetch, nodeMirrorBaseUrl)
+  if (!versionSpec) {
+    return allVersions.map(({ version }) => version)
+  }
+  if (versionSpec === 'latest') {
+    return [allVersions[0].version]
+  }
+  const { versions, versionRange } = filterVersions(allVersions, versionSpec)
+  return versions.filter(version => semver.satisfies(version, versionRange, SEMVER_OPTS))
+}
+
+async function fetchAllVersions (fetch: FetchFromRegistry, nodeMirrorBaseUrl?: string): Promise<NodeVersion[]> {
+  const response = await fetch(`${nodeMirrorBaseUrl ?? 'https://nodejs.org/download/release/'}index.json`)
+  return ((await response.json()) as NodeVersion[]).map(({ version, lts }) => ({
+    version: version.substring(1),
+    lts,
+  }))
 }
 
 function filterVersions (versions: NodeVersion[], versionSelector: string) {
   if (versionSelector === 'lts') {
     return {
-      versions: versions.filter(({ lts }) => lts !== false),
+      versions: versions
+        .filter(({ lts }) => lts !== false)
+        .map(({ version }) => version),
       versionRange: '*',
     }
   }
@@ -35,9 +62,14 @@ function filterVersions (versions: NodeVersion[], versionSelector: string) {
   if (vst?.type === 'tag') {
     const wantedLtsVersion = vst.normalized.toLowerCase()
     return {
-      versions: versions.filter(({ lts }) => typeof lts === 'string' && lts.toLowerCase() === wantedLtsVersion),
+      versions: versions
+        .filter(({ lts }) => typeof lts === 'string' && lts.toLowerCase() === wantedLtsVersion)
+        .map(({ version }) => version),
       versionRange: '*',
     }
   }
-  return { versions, versionRange: versionSelector }
+  return {
+    versions: versions.map(({ version }) => version),
+    versionRange: versionSelector,
+  }
 }