+++ packages/plugin-commands-deploy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/plugin-commands-deploy
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies [3dab7f83c]
+  - @pnpm/plugin-commands-installation@11.0.6
+  - @pnpm/cli-utils@1.0.6
+
 ## 2.0.5
 
 ### Patch Changes