+++ packages/plugin-commands-init/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/plugin-commands-init
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies [3dab7f83c]
+  - @pnpm/config@16.1.0
+  - @pnpm/cli-utils@1.0.6
+
 ## 2.0.5
 
 ### Patch Changes