+++ packages/plugin-commands-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/plugin-commands-server
 
+## 5.0.6
+
+### Patch Changes
+
+- Updated dependencies [3dab7f83c]
+  - @pnpm/config@16.1.0
+  - @pnpm/cli-utils@1.0.6
+  - @pnpm/store-connection-manager@5.0.6
+
 ## 5.0.5
 
 ### Patch Changes