+++ packages/store-connection-manager/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/store-connection-manager
 
+## 5.0.6
+
+### Patch Changes
+
+- Updated dependencies [3dab7f83c]
+  - @pnpm/config@16.1.0
+
 ## 5.0.5
 
 ### Patch Changes